修改DirectByteBufferPool相关的testcase
